Externalized OAuth classes into a submodule
Had to remove the code that was recently uncommented which adds parameters
to the url string. They are apparently set in the http request body.
Also had to modify the flow of SA_OAuthTwitterController's initWithEngine
method to allow the request token to be set before loading the web view
with the authorizeURLRequest.